+++ lib/src/clojure_lsp/db.clj
@@ -16,7 +16,7 @@
                  :dep-graph {}})
 (defonce db* (atom initial-db))
 
-(def version 11)
+(def version 12)
 
 (defn ^:private sqlite-db-file [project-root]
   (io/file (str project-root) ".lsp" ".cache" "sqlite.db"))